+++ public/js/app.js
@@ -2755,6 +2755,88 @@ __webpack_require__.r(__webpack_exports__);
 
 /***/ }),
 
+/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/StoryCard.vue?vue&type=script&lang=js&":
+/*!********************************************************************************************************************************************************************!*\
+  !*** ./node_modules/babel-loader/lib??ref--4-0!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/StoryCard.vue?vue&type=script&lang=js& ***!
+  \********************************************************************************************************************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+/* harmony default export */ __webpack_exports__["default"] = ({
+  props: {
+    single: {
+      type: Boolean,
+      required: false,
+      "default": false
+    },
+    story: {
+      type: Object,
+      required: true,
+      "default": function _default() {}
+    }
+  },
+  methods: {
+    remove: function remove() {
+      var _this = this;
+
+      axios["delete"](this.story.delete_url).then(function (response) {
+        _this.$toasted.success(_this.__("Story :name has been deleted", {
+          'name': _this.story.title
+        }));
+
+        if (_this.single) {
+          window.location = '/';
+        } else {
+          window.location.reload();
+        }
+      })["catch"](function (error) {
+        console.log(error);
+
+        _this.$toasted.error(_this.__("Unable to delete story"));
+      });
+    }
+  }
+});
+
+/***/ }),
+
 /***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/StoryForm.vue?vue&type=script&lang=js&":
 /*!********************************************************************************************************************************************************************!*\
   !*** ./node_modules/babel-loader/lib??ref--4-0!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/StoryForm.vue?vue&type=script&lang=js& ***!
@@ -2766,6 +2848,7 @@ __webpack_require__.r(__webpack_exports__);
 __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var mavon_editor__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! mavon-editor */ "./node_modules/mavon-editor/dist/mavon-editor.js");
 /* harmony import */ var mavon_editor__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(mavon_editor__WEBPACK_IMPORTED_MODULE_0__);
+//
 //
 //
 //
@@ -2823,15 +2906,6 @@ Vue.use(mavon_editor__WEBPACK_IMPORTED_MODULE_0___default.a);
     Editor: mavon_editor__WEBPACK_IMPORTED_MODULE_0___default.a
   },
   props: {
-    submitUrl: {
-      type: String,
-      required: true
-    },
-    method: {
-      type: String,
-      required: false,
-      "default": 'POST'
-    },
     story: {
       type: Object,
       required: false,
@@ -2886,8 +2960,8 @@ Vue.use(mavon_editor__WEBPACK_IMPORTED_MODULE_0___default.a);
 
       this.loading = true;
       axios.request({
-        method: this.method,
-        url: this.submitUrl,
+        method: this.story ? 'PUT' : 'POST',
+        url: this.story ? this.story.update_url : '/api/story',
         data: this.form
       }).then(function (response) {
         if (_this.story) {
@@ -66123,6 +66197,149 @@ render._withStripped = true
 
 /***/ }),
 
+/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/StoryCard.vue?vue&type=template&id=3ea8b54a&":
+/*!************************************************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/StoryCard.vue?vue&type=template&id=3ea8b54a& ***!
+  \************************************************************************************************************************************************************************************************************/
+/*! exports provided: render, staticRenderFns */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "render", function() { return render; });
+/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return staticRenderFns; });
+var render = function() {
+  var _vm = this
+  var _h = _vm.$createElement
+  var _c = _vm._self._c || _h
+  return _c(
+    "div",
+    {
+      staticClass: "card card--chest mb-4",
+      class: { "card-single": _vm.single, "card-index": !_vm.single }
+    },
+    [
+      _c("div", { staticClass: "card-body" }, [
+        _c("h5", { staticClass: "card-title" }, [
+          _c("span", [_vm._v(_vm._s(_vm.__("Story")))]),
+          _vm._v(" — "),
+          _c("a", { attrs: { href: _vm.story.url } }, [
+            _vm._v(_vm._s(_vm.story.title))
+          ])
+        ]),
+        _vm._v(" "),
+        _c(
+          "div",
+          { staticClass: "card-text card-reduce" },
+          [_c("vue-markdown", [_vm._v(_vm._s(_vm.story.content))])],
+          1
+        ),
+        _vm._v(" "),
+        _vm.story.tags.length > 0
+          ? _c(
+              "p",
+              { staticClass: "card-text mt-1" },
+              _vm._l(_vm.story.tags, function(tag) {
+                return _c(
+                  "a",
+                  {
+                    staticClass: "badge badge-secondary",
+                    attrs: { href: "/tag/" + tag }
+                  },
+                  [_vm._v(_vm._s(tag))]
+                )
+              }),
+              0
+            )
+          : _vm._e()
+      ]),
+      _vm._v(" "),
+      _c("div", { staticClass: "card-footer d-flex justify-content-between" }, [
+        _c("span", [
+          _vm._v(
+            _vm._s(_vm.story.is_private ? "🔒 " : "") +
+              _vm._s(_vm.story.date_formated)
+          )
+        ]),
+        _vm._v(" "),
+        _c("div", { staticClass: "dropdown" }, [
+          _c(
+            "button",
+            {
+              staticClass: "btn btn-outline-dark btn-sm dropdown-toggle",
+              attrs: {
+                type: "button",
+                id: "dropdownMenuButton",
+                "data-toggle": "dropdown",
+                "aria-haspopup": "true",
+                "aria-expanded": "false"
+              }
+            },
+            [
+              _vm._v(
+                "\n                " + _vm._s(_vm.__("More")) + "\n            "
+              )
+            ]
+          ),
+          _vm._v(" "),
+          _c(
+            "div",
+            {
+              staticClass: "dropdown-menu dropdown-menu-right",
+              attrs: { "aria-labelledby": "dropdownMenuButton" }
+            },
+            [
+              _c(
+                "a",
+                {
+                  staticClass: "dropdown-item",
+                  attrs: { href: _vm.story.url }
+                },
+                [_vm._v(_vm._s(_vm.__("Permalink")))]
+              ),
+              _vm._v(" "),
+              _vm.story.editable
+                ? _c("h6", { staticClass: "dropdown-header" }, [
+                    _vm._v(_vm._s(_vm.__("Manage")))
+                  ])
+                : _vm._e(),
+              _vm._v(" "),
+              _vm.story.editable
+                ? _c(
+                    "a",
+                    {
+                      staticClass: "dropdown-item",
+                      attrs: { href: _vm.story.edit_url }
+                    },
+                    [_vm._v(_vm._s(_vm.__("Edit")))]
+                  )
+                : _vm._e(),
+              _vm._v(" "),
+              _vm.story.editable
+                ? _c("confirm", {
+                    staticClass: "dropdown-item",
+                    attrs: {
+                      text: _vm.__("Delete"),
+                      "text-confirm": _vm.__("Confirm")
+                    },
+                    on: { confirmed: _vm.remove }
+                  })
+                : _vm._e()
+            ],
+            1
+          )
+        ])
+      ])
+    ]
+  )
+}
+var staticRenderFns = []
+render._withStripped = true
+
+
+
+/***/ }),
+
 /***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/StoryForm.vue?vue&type=template&id=5b624cfe&":
 /*!************************************************************************************************************************************************************************************************************!*\
   !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/StoryForm.vue?vue&type=template&id=5b624cfe& ***!
@@ -66300,7 +66517,18 @@ var render = function() {
             : _vm._e(),
           _vm._v("\n            " + _vm._s(_vm.__("Save")) + "\n        ")
         ]
-      )
+      ),
+      _vm._v(" "),
+      _vm.story
+        ? _c(
+            "a",
+            {
+              staticClass: "btn btn-outline-primary",
+              attrs: { href: _vm.story.url }
+            },
+            [_vm._v(_vm._s(_vm.__("View")))]
+          )
+        : _vm._e()
     ])
   ])
 }
@@ -82265,6 +82493,7 @@ var map = {
 	"./components/ManageTags.vue": "./resources/js/components/ManageTags.vue",
 	"./components/Search.vue": "./resources/js/components/Search.vue",
 	"./components/Sharer.vue": "./resources/js/components/Sharer.vue",
+	"./components/StoryCard.vue": "./resources/js/components/StoryCard.vue",
 	"./components/StoryForm.vue": "./resources/js/components/StoryForm.vue",
 	"./components/Tags.vue": "./resources/js/components/Tags.vue"
 };
@@ -83071,6 +83300,75 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "render", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_Sharer_vue_vue_type_template_id_10eae91e___WEBPACK_IMPORTED_MODULE_0__["render"]; });
 
 /* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_Sharer_vue_vue_type_template_id_10eae91e___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"]; });
+
+
+
+/***/ }),
+
+/***/ "./resources/js/components/StoryCard.vue":
+/*!***********************************************!*\
+  !*** ./resources/js/components/StoryCard.vue ***!
+  \***********************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony import */ var _StoryCard_vue_vue_type_template_id_3ea8b54a___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./StoryCard.vue?vue&type=template&id=3ea8b54a& */ "./resources/js/components/StoryCard.vue?vue&type=template&id=3ea8b54a&");
+/* harmony import */ var _StoryCard_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./StoryCard.vue?vue&type=script&lang=js& */ "./resources/js/components/StoryCard.vue?vue&type=script&lang=js&");
+/* empty/unused harmony star reexport *//* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ "./node_modules/vue-loader/lib/runtime/componentNormalizer.js");
+
+
+
+
+
+/* normalize component */
+
+var component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__["default"])(
+  _StoryCard_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__["default"],
+  _StoryCard_vue_vue_type_template_id_3ea8b54a___WEBPACK_IMPORTED_MODULE_0__["render"],
+  _StoryCard_vue_vue_type_template_id_3ea8b54a___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"],
+  false,
+  null,
+  null,
+  null
+  
+)
+
+/* hot reload */
+if (false) { var api; }
+component.options.__file = "resources/js/components/StoryCard.vue"
+/* harmony default export */ __webpack_exports__["default"] = (component.exports);
+
+/***/ }),
+
+/***/ "./resources/js/components/StoryCard.vue?vue&type=script&lang=js&":
+/*!************************************************************************!*\
+  !*** ./resources/js/components/StoryCard.vue?vue&type=script&lang=js& ***!
+  \************************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_StoryCard_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/babel-loader/lib??ref--4-0!../../../node_modules/vue-loader/lib??vue-loader-options!./StoryCard.vue?vue&type=script&lang=js& */ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/StoryCard.vue?vue&type=script&lang=js&");
+/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__["default"] = (_node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_StoryCard_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__["default"]); 
+
+/***/ }),
+
+/***/ "./resources/js/components/StoryCard.vue?vue&type=template&id=3ea8b54a&":
+/*!******************************************************************************!*\
+  !*** ./resources/js/components/StoryCard.vue?vue&type=template&id=3ea8b54a& ***!
+  \******************************************************************************/
+/*! exports provided: render, staticRenderFns */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_StoryCard_vue_vue_type_template_id_3ea8b54a___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../node_modules/vue-loader/lib??vue-loader-options!./StoryCard.vue?vue&type=template&id=3ea8b54a& */ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/StoryCard.vue?vue&type=template&id=3ea8b54a&");
+/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "render", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_StoryCard_vue_vue_type_template_id_3ea8b54a___WEBPACK_IMPORTED_MODULE_0__["render"]; });
+
+/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_StoryCard_vue_vue_type_template_id_3ea8b54a___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"]; });