+++ resources/views/chest.blade.php
@@ -10,9 +10,7 @@
 <div class="container">
     <div class="row justify-content-center">
         <div class="col-12 col-md-8">
-            <chest-card :single="true"
-                        :chest="{{ json_encode(\App\Http\Resources\ChestResource::make($chest)) }}"
-            ></chest-card>
+            <chest-card :single="true" :chest="{{ json_encode(\App\Http\Resources\ChestResource::make($chest)) }}"></chest-card>
         </div>
     </div>
 </div>