+++ resources/views/tag.blade.php
@@ -16,7 +16,7 @@
                     @if($post->postable instanceof \App\Link)
                         @include('partials.link', ['link' => $post->postable, 'post' => $post, 'index' => true])
                     @elseif($post->postable instanceof \App\Story)
-                        @include('partials.story', ['story' => $post->postable, 'post' => $post, 'index' => true])
+                        <story-card :single="false" :story="{{ json_encode(\App\Http\Resources\StoryResource::make($post->postable)) }}"></story-card>
                     @elseif($post->postable instanceof \App\Chest)
                         <chest-card :single="false" :chest="{{ json_encode(\App\Http\Resources\ChestResource::make($post->postable)) }}"></chest-card>
                     @endif