+++ routes/api.php
@@ -13,6 +13,7 @@ Route::put('link/{id}', 'LinkController@update')->name('link.update');
 
 Route::post('story', 'StoryController@store')->name('story.store');
 Route::put('story/{id}', 'StoryController@update')->name('story.update');
+Route::delete('story/{id}', 'StoryController@delete')->name('story.delete');
 
 Route::post('chest', 'ChestController@store')->name('chest.store');
 Route::put('chest/{id}', 'ChestController@update')->name('chest.update');